Feat: 팀 기능 구현 및 S3 세팅

+++ src/main/java/com/whylog/server/domain/decision/dto/DecisionResponse.java
@@ -22,7 +22,7 @@ public class DecisionResponse {
         private String name;
 
         @Schema(description = "적용사항 개수", example = "3")
-        private Long applicationCount;
+        private Integer applicationCount;
     }
 
     @Getter
